refactor: move COSEKey record
