+++ src/main/java/uk/gov/di/mobile/wallet/cri/credential/mobile_driving_licence/cose/COSEKey.java
@@ -1,4 +1,4 @@
-package uk.gov.di.mobile.wallet.cri.credential.mobile_driving_licence.mdoc;
+package uk.gov.di.mobile.wallet.cri.credential.mobile_driving_licence.cose;
 
 import java.util.Map;
