+++ src/main/java/uk/gov/di/mobile/wallet/cri/credential/mobile_driving_licence/mdoc/DeviceKeyInfo.java
@@ -1,5 +1,7 @@
 package uk.gov.di.mobile.wallet.cri.credential.mobile_driving_licence.mdoc;
 
+import uk.gov.di.mobile.wallet.cri.credential.mobile_driving_licence.cose.COSEKey;
+
 public record DeviceKeyInfo(COSEKey deviceKey, KeyAuthorizations keyAuthorizations
-        //        KeyInfo keyInfo
+                            //        KeyInfo keyInfo
         ) {}
